refactor(zwave_js): isolate the UC fallback into a removable support layer
Move every User Code CC fallback field, helper, and push handler into
ZWaveJSUserCodeFallbackSupport (providers/_zwave_js_uc.py), an
intermediate class between BaseLock and ZWaveJSLock, and consolidate all
fallback tests into tests/providers/zwave_js/test_uc_fallback.py. The
concrete provider keeps only the routing branch points.

Once the upstream driver fix (zwave-js/zwave-js#8873) is the minimum
supported version, the workaround can be deleted as a unit: drop the
module, re-base ZWaveJSLock on BaseLock, remove the branch points, and
delete the test file plus its conftest fixtures. The module docstring
carries the full removal recipe.

Entire-Checkpoint: 91cbca6795ab
