+++ tests/providers/zwave_js/test_e2e.py
@@ -27,8 +27,6 @@ from custom_components.lock_code_manager.const import (
 from custom_components.lock_code_manager.domain.models import SlotCredential
 from custom_components.lock_code_manager.providers.zwave_js import ZWaveJSLock
 
-from .conftest import get_zwave_lock, uc_only_caps_response, uc_slot_walk
-
 
 def async_capture_events(
     hass: HomeAssistant, event_name: str
@@ -274,64 +272,3 @@ class TestEvents:
         # Credential events push unreadable (the lock doesn't expose the Personal
         # Identification Number value in the event; a coordinator refresh reads it).
         assert e2e_zwave_lock.coordinator.data.get(1) == SlotCredential.unreadable()
-
-
-class TestUCFallbackLifecycle:
-    """Full LCM lifecycle against a lock in User Code CC fallback mode (issue #1251)."""
-
-    async def test_setup_succeeds_and_writes_route_through_uc_utils(
-        self,
-        hass: HomeAssistant,
-        zwave_integration: MockConfigEntry,
-        lock_entity: er.RegistryEntry,
-        mock_access_control: MagicMock,
-        mock_lock_helpers: dict,
-        mock_uc_utils: dict,
-    ) -> None:
-        """A UC-fallback lock completes LCM setup and routes PIN writes via UC utils.
-
-        This pins two regressions at once: setup must not reject the
-        slot-only capabilities the fallback reports (the base used to
-        require ``supports_user_management``), and writes must use the
-        User Code CC utilities instead of the unified API whose broken
-        capability data caused 'between 1 and 0' rejections.
-        """
-        mock_lock_helpers[
-            "async_get_credential_capabilities"
-        ].return_value = uc_only_caps_response()
-        mock_uc_utils["get_usercodes"].return_value = uc_slot_walk(
-            30, occupied={2: "1234"}
-        )
-
-        lcm_entry = MockConfigEntry(
-            domain=DOMAIN,
-            data={
-                CONF_LOCKS: [lock_entity.entity_id],
-                CONF_SLOTS: {"2": {}, "4": {}},
-            },
-            unique_id="test_zwave_js_uc_e2e",
-        )
-        lcm_entry.add_to_hass(hass)
-        assert await hass.config_entries.async_setup(lcm_entry.entry_id)
-        await hass.async_block_till_done()
-
-        try:
-            lock = get_zwave_lock(hass, lcm_entry, lock_entity)
-            lock._min_operation_delay = 0.0
-
-            # Setting a code skips the user lifecycle and writes via set_usercode
-            result = await lock.async_set_usercode(4, "5678", "Test User")
-            assert result is True
-            mock_lock_helpers["async_set_user"].assert_not_called()
-            mock_uc_utils["set_usercode"].assert_awaited_once_with(lock.node, 4, "5678")
-            mock_access_control.set_credential.assert_not_called()
-
-            # Clearing resolves the owner from the synthesized UC users and
-            # clears via clear_usercode
-            result = await lock.async_clear_usercode(2)
-            assert result is True
-            mock_uc_utils["clear_usercode"].assert_awaited_once_with(lock.node, 2)
-            mock_access_control.delete_credential.assert_not_called()
-            mock_lock_helpers["async_delete_credential"].assert_not_called()
-        finally:
-            await hass.config_entries.async_unload(lcm_entry.entry_id)
