+++ tests/providers/zwave_js/test_events.py
@@ -6,9 +6,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 from pytest_homeassistant_custom_component.common import MockConfigEntry
-from zwave_js_server.const import CommandClass
 from zwave_js_server.const.command_class.access_control import UserCredentialType
-from zwave_js_server.const.command_class.lock import CodeSlotStatus
 from zwave_js_server.event import Event as ZwaveEvent
 from zwave_js_server.model.node import Node
 
@@ -601,307 +599,3 @@ async def test_credential_deleted_non_pin_ignored(
     mock_coordinator.push_update.assert_not_called()
 
     zwave_js_lock.unsubscribe_push_updates()
-
-
-# ---------------------------------------------------------------------------
-# User Code CC fallback push handling (issue #1251)
-# ---------------------------------------------------------------------------
-
-
-def _make_uc_value_event(
-    node_id: int, property_name: str, code_slot: int, new_value
-) -> ZwaveEvent:
-    """Create a User Code CC value-updated ZwaveEvent."""
-    return ZwaveEvent(
-        type="value updated",
-        data={
-            "source": "node",
-            "event": "value updated",
-            "nodeId": node_id,
-            "args": {
-                "commandClass": CommandClass.USER_CODE,
-                "property": property_name,
-                "propertyKey": code_slot,
-                "newValue": new_value,
-            },
-        },
-    )
-
-
-def _make_duplicate_code_event(node_id: int, user_id: int | None = None) -> ZwaveEvent:
-    """Create a duplicate code notification ZwaveEvent."""
-    params: dict = {}
-    if user_id is not None:
-        params["userId"] = user_id
-    return ZwaveEvent(
-        type="notification",
-        data={
-            "source": "node",
-            "event": "notification",
-            "nodeId": node_id,
-            "endpointIndex": 0,
-            "ccId": 113,
-            "args": {
-                "type": 6,  # ACCESS_CONTROL
-                "event": 15,  # NEW_USER_CODE_NOT_ADDED_DUE_TO_DUPLICATE_CODE
-                "label": "Access Control",
-                "eventLabel": "New user code not added due to duplicate code",
-                "parameters": params,
-            },
-        },
-    )
-
-
-async def test_subscribe_uc_mode_registers_value_listener_only(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-) -> None:
-    """In UC-fallback mode, push uses one value-updated listener."""
-    lcm_entry = MockConfigEntry(domain=DOMAIN, data={CONF_LOCKS: [], CONF_SLOTS: {}})
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    uc_fallback_lock.subscribe_push_updates()
-    assert len(uc_fallback_lock._push_unsubs) == 1
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    await uc_fallback_lock.async_unload(False)
-
-
-async def test_subscribe_unknown_mode_registers_both_listener_sets(
-    hass: HomeAssistant,
-    zwave_js_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    mock_access_control: MagicMock,
-    mock_lock_helpers: dict,
-) -> None:
-    """Before the capability probe runs, both listener sets are registered.
-
-    The handlers are self-filtering and pushes are idempotent, so
-    double-coverage is safe; it only happens when push subscription
-    races ahead of the first capability probe.
-    """
-    zwave_js_lock.subscribe_push_updates()
-    assert len(zwave_js_lock._push_unsubs) == 4
-
-    zwave_js_lock.unsubscribe_push_updates()
-
-
-async def test_uc_push_masked_code_sends_unreadable(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-    mock_uc_utils: dict,
-) -> None:
-    """A masked userCode value update pushes SlotCredential.unreadable()."""
-    lcm_entry = MockConfigEntry(
-        domain=DOMAIN,
-        data={
-            CONF_LOCKS: [uc_fallback_lock.lock.entity_id],
-            CONF_SLOTS: {"2": {}},
-        },
-    )
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    mock_coordinator = MagicMock()
-    mock_coordinator.data = {}
-    uc_fallback_lock.coordinator = mock_coordinator
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    uc_fallback_lock.subscribe_push_updates()
-
-    lock_schlage_be469.receive_event(
-        _make_uc_value_event(lock_schlage_be469.node_id, "userCode", 2, "****")
-    )
-    await hass.async_block_till_done()
-
-    mock_coordinator.push_update.assert_called_once_with(
-        {2: SlotCredential.unreadable()}
-    )
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    await uc_fallback_lock.async_unload(False)
-
-
-async def test_uc_push_status_available_ignored_when_pin_expected(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-    mock_uc_utils: dict,
-) -> None:
-    """A stale AVAILABLE status is ignored when LCM expects a PIN on the slot.
-
-    Some locks send stale AVAILABLE events after a code was set; acting
-    on them would cause infinite sync loops.
-    """
-    lcm_entry = MockConfigEntry(
-        domain=DOMAIN,
-        data={
-            CONF_LOCKS: [uc_fallback_lock.lock.entity_id],
-            CONF_SLOTS: {"2": {}},
-        },
-    )
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    mock_coordinator = MagicMock()
-    mock_coordinator.data = {2: SlotCredential.known("1234")}
-    mock_coordinator.desired_credential.return_value = SlotCredential.known("1234")
-    uc_fallback_lock.coordinator = mock_coordinator
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    uc_fallback_lock.subscribe_push_updates()
-
-    lock_schlage_be469.receive_event(
-        _make_uc_value_event(
-            lock_schlage_be469.node_id, "userIdStatus", 2, CodeSlotStatus.AVAILABLE
-        )
-    )
-    await hass.async_block_till_done()
-
-    mock_coordinator.push_update.assert_not_called()
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    await uc_fallback_lock.async_unload(False)
-
-
-async def test_uc_push_status_available_clears_slot_when_no_pin_expected(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-    mock_uc_utils: dict,
-) -> None:
-    """An AVAILABLE status pushes empty when LCM does not expect a PIN."""
-    lcm_entry = MockConfigEntry(
-        domain=DOMAIN,
-        data={
-            CONF_LOCKS: [uc_fallback_lock.lock.entity_id],
-            CONF_SLOTS: {"2": {}},
-        },
-    )
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    mock_coordinator = MagicMock()
-    mock_coordinator.data = {2: SlotCredential.known("1234")}
-    mock_coordinator.desired_credential.return_value = SlotCredential.empty()
-    uc_fallback_lock.coordinator = mock_coordinator
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    uc_fallback_lock.subscribe_push_updates()
-
-    lock_schlage_be469.receive_event(
-        _make_uc_value_event(
-            lock_schlage_be469.node_id, "userIdStatus", 2, CodeSlotStatus.AVAILABLE
-        )
-    )
-    await hass.async_block_till_done()
-
-    mock_coordinator.push_update.assert_called_once_with({2: SlotCredential.empty()})
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    await uc_fallback_lock.async_unload(False)
-
-
-async def test_uc_duplicate_code_notification_marks_rejected(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-    mock_uc_utils: dict,
-) -> None:
-    """The duplicate-code notification marks the in-flight slot rejected."""
-    lcm_entry = MockConfigEntry(domain=DOMAIN, data={CONF_LOCKS: [], CONF_SLOTS: {}})
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    uc_fallback_lock._set_in_progress_code_slot = 2
-    lock_schlage_be469.receive_event(
-        _make_duplicate_code_event(lock_schlage_be469.node_id, user_id=2)
-    )
-    await hass.async_block_till_done()
-
-    assert 2 in uc_fallback_lock._rejected_code_slots
-    assert uc_fallback_lock._set_in_progress_code_slot is None
-
-    await uc_fallback_lock.async_unload(False)
-
-
-async def test_uc_duplicate_code_notification_user_id_zero_marks_in_flight_slot(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-    mock_uc_utils: dict,
-) -> None:
-    """Firmwares reporting userId=0 still reject the slot LCM is setting."""
-    lcm_entry = MockConfigEntry(domain=DOMAIN, data={CONF_LOCKS: [], CONF_SLOTS: {}})
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    uc_fallback_lock._set_in_progress_code_slot = 3
-    lock_schlage_be469.receive_event(
-        _make_duplicate_code_event(lock_schlage_be469.node_id, user_id=0)
-    )
-    await hass.async_block_till_done()
-
-    assert 3 in uc_fallback_lock._rejected_code_slots
-    assert uc_fallback_lock._set_in_progress_code_slot is None
-
-    await uc_fallback_lock.async_unload(False)
-
-
-async def test_uc_duplicate_code_notification_ignored_when_not_in_progress(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-    mock_uc_utils: dict,
-) -> None:
-    """Without an in-flight set, the notification is not attributed to any slot."""
-    lcm_entry = MockConfigEntry(domain=DOMAIN, data={CONF_LOCKS: [], CONF_SLOTS: {}})
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    lock_schlage_be469.receive_event(
-        _make_duplicate_code_event(lock_schlage_be469.node_id, user_id=2)
-    )
-    await hass.async_block_till_done()
-
-    assert not uc_fallback_lock._rejected_code_slots
-
-    await uc_fallback_lock.async_unload(False)
-
-
-async def test_uc_set_in_progress_cleared_on_value_update(
-    hass: HomeAssistant,
-    uc_fallback_lock: ZWaveJSLock,
-    zwave_integration: MockConfigEntry,
-    lock_schlage_be469: Node,
-    mock_uc_utils: dict,
-) -> None:
-    """A userCode update for the in-flight slot closes the rejection window."""
-    lcm_entry = MockConfigEntry(domain=DOMAIN, data={CONF_LOCKS: [], CONF_SLOTS: {}})
-    lcm_entry.add_to_hass(hass)
-    await uc_fallback_lock.async_setup_internal(lcm_entry)
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    uc_fallback_lock.subscribe_push_updates()
-
-    uc_fallback_lock._set_in_progress_code_slot = 2
-    lock_schlage_be469.receive_event(
-        _make_uc_value_event(lock_schlage_be469.node_id, "userCode", 2, "5678")
-    )
-    await hass.async_block_till_done()
-
-    assert uc_fallback_lock._set_in_progress_code_slot is None
-
-    uc_fallback_lock.unsubscribe_push_updates()
-    await uc_fallback_lock.async_unload(False)
